requirements.txt
pytest
pytest-playwright
playwright
uuid
requests
pytest-html
allure-pytest
behave
allure-behave